Allow separate control message bus
ombt2 relies on the message bus to control the servers and clients
that run the tests.  This may be undesireable if the test is stressful
or destructive to the message bus.

This change allows ombt2 to use a second message bus just for test
control traffic.  Test are not run against the control message bus

Closes #2